Fix no updates happening in real component
How
Added a trackedNested decorator so we can easily notify invoking class
of updates.

We now retain a copy of the unproxied data because ember tracked does
not work well with proxy. So all update requests are captured by proxy
and then sent to raw data for updates. We then nuke the original proxied
data and reproxy a clone of the raw data. Then we use ember set to
notify the class member of the change.