Fix trackedNested not working with initial null assignment and property
Tracked nested decorator tried to call the property that is being evaled
with the correct context but if it's initially not executable it will
return an error.

Also for the optional chaining operators where propties can be null, we
added a check where the object must be an array or object or ember
object, as to not to try to proxify null or access properties of null.