Revamped decorator to re-proxy value when class property reassigned
Current we call the decorator to proxy the initial value. However,
subsequent reassignment will not be decorated. i.e this.a = null and
then you set this.a = { b: 1 }

This change will make sure we also create a setter for the key in the
class so that any subsequent changes to the property will also proxy the
property with trackedNested.